Updated DamageCal
I made it so that it returned the double value of damage. Also changed some of the data type into double because there might be decimals in the stats after checking back. Please refer to this one when you try to shrink the program later, sorry if you already have.